requirements.txt
# Core LangChain Libraries - Let pip resolve versions
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.3.0

# LangServe and its dependencies
langserve>=0.3.0
fastapi>=0.110.0
uvicorn>=0.30.0
sse-starlette>=1.6.0
pydantic>=2.7.0

# Other utilities
ollama>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.35.0